# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Empresas",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Entorno de Producción (Consume Créditos)"
        },
        {
            "url": "/api/sandbox/v1",
            "description": "Entorno Sandbox de Pruebas (Gratis. Usa CIF: A15075062)"
        }
    ],
    "paths": {
...
        "/api/v1/companies/score": {
            "get": {
                "tags": [
                    "2. Plan Pro"
                ],
                "summary": "Scoring Comercial",
                "description": "Obtener la puntuación y desglose de scoring para una empresa. **Coste:** 1 llamada de tu cuota mensual (plan suscripción) o 3 créditos del monedero (bono prepago). Las respuestas con error (400, 404, 403, etc.) no consumen cuota ni créditos.",
                "operationId": "9b4661101275cf401d165674ea348e66",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "query",
                        "description": "CIF de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del scoring comercial",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/companies/signals": {
            "get": {
                "tags": [
                    "2. Plan Pro"
                ],
                "summary": "Señales Societarias",
                "description": "Obtener eventos societarios relevantes (BORME) de una empresa. **Coste:** 1 llamada de tu cuota mensual (plan suscripción) o 3 créditos del monedero (bono prepago). Las respuestas con error (400, 404, 403, etc.) no consumen cuota ni créditos.",
                "operationId": "03e9cb58fe9920dd2a821da49ddd7921",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "query",
                        "description": "CIF de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Señales societarias de la empresa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/companies/insights": {
            "get": {
                "tags": [
                    "3. Plan Business"
                ],
                "summary": "Análisis con Inteligencia Artificial",
                "description": "Obtener un análisis profundo y perfilado de ventas mediante IA. **Coste:** 1 llamada de tu cuota mensual (plan suscripción) o 3 créditos del monedero (bono prepago). Las respuestas con error (400, 404, 403, etc.) no consumen cuota ni créditos.",
                "operationId": "18786d5e2d8d8be91882f39936c5756d",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "query",
                        "description": "CIF de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insights generados por IA",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/companies/contact-prep": {
            "get": {
                "tags": [
                    "3. Plan Business"
                ],
                "summary": "Preparación de Contacto",
                "description": "Obtener tácticas recomendadas por la IA para contactar o vender a la empresa. **Coste:** 1 llamada de tu cuota mensual (plan suscripción) o 3 créditos del monedero (bono prepago). Las respuestas con error (400, 404, 403, etc.) no consumen cuota ni créditos.",
                "operationId": "85cf72fa9acdc26beeb7147ec1e768b3",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "query",
                        "description": "CIF de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tácticas de preparación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Introduce tu API Key privada aquí para autenticarte y poder probar los endpoints en tiempo real.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}